Update Azure DevOps workflow
This commit updates the Azure DevOps workflow by replacing the deprecated action `microsoft/azure-devops-extension-tasks/run@v1` with `actions/setup-node@v2` and sets the node version to 16. This change ensures compatibility with the latest version of Node.js and improves the overall workflow. The `run` command has also been adjusted to use `npm` before executing the `tfx extension publish` command with the appropriate arguments. These modifications aim to enhance the publishing process to the Azure DevOps Marketplace and ensure a seamless experience for users.