Release extension to Azure DevOps Marketplace
This commit adds a modification to the workflow file `.github/workflows/release_to_azure_devops.yaml`. It includes the addition of a command to change the directory to `src` before publishing the extension to the Azure DevOps Marketplace. This change ensures that the extension is published from the correct location.