Update release action in workflow
Updating the release action in the workflow for publishing to Azure DevOps Marketplace. The action now includes the flag "--token" to utilize the Azure DevOps PAT secret. Additionally, the command to retrieve the extension version now includes the token as well. The TFS_PAT environment variable is no longer used. This change ensures the extension is published with the correct permissions and the extension version is properly retrieved.