Update release_to_azure_devops.yaml workflow
This commit updates the release_to_azure_devops.yaml workflow. It modifies the `tfx extension publish` command by adding the `--rev-version` flag. This flag ensures that the extension version is revised before publishing it to the Azure DevOps Marketplace. The extension version is retrieved using the `tfx extension show` command and saved as an environment variable `EXTENSION_VERSION`.