Publish extension version in GitHub release action
Publishes the extension version to the Azure DevOps Marketplace using the `tfx extension publish` command. The extension version is obtained using the `tfx extension show` command. The obtained version is then passed as the value for the `release_name` input in the GitHub Release action. This ensures that the release name is set to the current extension version.