Update task version script modified to accept a semver argument
The update-task-version.js script has been modified to accept a semver argument instead of the path to the vss-extension.json file. The previous usage error message has been updated accordingly. The script now takes the semver argument and updates the task.json file with the provided version. The task version update is logged to the console.