Update Azure DevOps extension version in release workflow
This commit updates the release workflow for an Azure DevOps extension by upgrading the version number used in the `tfx extension create` command. The previous command simply used the `--extension-version` flag, but now it includes an `--override` flag with a JSON object containing the version. This change ensures that the correct extension version is created and published to the Azure DevOps Marketplace.