Fix versioning in release workflow
Updated the release workflow file in order to fix the versioning logic. Previously, the version number was created by appending the GITHUB_RUN_ID to the major, next minor, and patch numbers. However, it was discovered that the GITHUB_RUN_ID is too long for the version number format. Therefore, a new logic was introduced to calculate a shorter version number by taking the modulo of GITHUB_RUN_ID by 1000000000. This ensures that the version number remains within the expected format and avoids any potential issues.