Add status check for formatting errors in PRs
- Add a flag `statusCheck` to enable/disable setting a status check in PR when formatting errors are found. Default value is `false`.
- Add an input `statusCheckName` to specify the name of the status check in the PR for formatting errors. Default value is `Code Format`.

This change adds the ability to set a status check in the PR when formatting errors are found. If `statusCheck` is set to `true`, the task will create a status check with the name specified in `statusCheckName`. This can be used to track and enforce code formatting standards in pull requests.