Publish the Azure DevOps extension with the new version number
This commit updates the release workflow in `.github/workflows/release_to_azure_devops.yaml`. It adds conditional statements to determine whether the event is a pull request. If it is, the extension will be created with the version specified in the environment variable `EXT_VERSION` and marked as not public. If it is not a pull request, the extension will be created with the version specified in `EXT_VERSION` and marked as public.

The extension is then published to the Azure DevOps Marketplace using the `tfx extension publish` command. If it is a pull request, the extension will be published to the organization specified in `AZURE_DEVOPS_ORGANIZATION` and not made public. Otherwise, it will be published and made public.