Update Azure DevOps extension name in release workflow
The `release_to_azure_devops.yaml` workflow file has been modified to update the name of the Azure DevOps extension that is being published. Previously, it was publishing the extension with the name `ajeckmans.format-check`, but now it has been changed to `ajeckmans.format-check-private`. This change ensures that the correct extension is being published to the Azure DevOps organization.