Update format-check.ts and task.json
In format-check.ts, modify the initialization of the statusCheckContext object by removing the default values for the name and genre properties. Now, the values will be taken directly from the corresponding environment variables.

In task.json, update the defaultValue of the statusCheckName property to "format check" and add a new property called statusCheckGenre with a defaultValue of "formatting". These values will be used for the status check in the PR for formatting errors when the 'Set PR Status Check' is true.