Update task version in `.github/workflows/update-task-version.js`
This commit modifies the `.github/workflows/update-task-version.js` file to update the task version. It updates the `Patch` value in the `version` object by combining the third and fourth elements of the `version` array. The updated value is calculated as `version[2] * 1000000 + version[3]`. This change ensures the task version is properly updated in the JSON file.