Fix setting of gitStatusState in setPRStatusAndFailTask
Setting the gitStatusState in `setPRStatusAndFailTask` was incorrect. Previously, the status was set to `Failed` even if `formatIssuesExist` was false and `taskParams.failOnFormattingErrors` was also false. This meant that the status would always be set to `Failed` even if there were no formatting errors.

This commit fixes the logic and now sets the gitStatusState to `Failed` only if `formatIssuesExist` is true and `taskParams.failOnFormattingErrors` is also true.