fix: Fix issue with closing resolved comment threads
This commit fixes an issue in the `markResolvedThreadsAsClosed` function where closed comment threads were not being skipped. Now, if a thread is already closed, it will be skipped in the processing loop.

In addition, this commit adds logic to update the pull request status based on the result of the code format check. If there are formatting issues and the `failOnFormattingErrors` flag is set, the pull request status will be set to "Failed". If there are no formatting issues or the `failOnFormattingErrors` flag is not set, the pull request status will be set to "Succeeded".

Note: This commit message excludes changes to imports and requires.