Refactor: Modify AnnotatedReport and FormatReport classes
The AnnotatedReport class has been modified to include a new constructor that takes commitId and changeType parameters. It sets the commitId and changeType properties of the class.

The FormatReport class has been modified to use the definite assignment assertion syntax (!) for the DocumentId, FileName, FilePath, and FileChanges properties. This indicates that these properties will always be assigned a value.

These changes improve the clarity and maintainability of the code.