Fix error in `getPullRequestChanges` method
In the `PullRequestService` class, the `getPullRequestChanges` method has been updated to fix an error.

Previously, the method was making a call to `gitApi.getRefs` using the `pr.targetRefName` directly. However, this caused an issue when the `targetRefName` contained the prefix "refs/". Now, before making the call, the `targetRefName` is sanitized by removing the "refs/" prefix. This ensures the correct value is passed to `gitApi.getRefs`.

If the `targetRefs` array is empty, it means that the target ref was not found. In this case, an error message is logged with the sanitized target ref name.

These changes address the issue and improve the reliability of the `getPullRequestChanges` method.