Update git commit diffs parameters
This commit updates the parameters for retrieving commit diffs in the `PullRequestService` class. The `baseVersion` parameter has been changed to `version`, and the `baseVersionType` parameter has been changed to `versionType`, while the `targetVersion` parameter has been changed to `version` and the `targetVersionType` parameter has been changed to `versionType`. Additionally, the value of the `baseVersion` parameter has been updated to `targetBranch`, and the value of the `targetVersion` parameter has been updated to `this.settings.Environment.pullRequestSourceCommit`.