chore: Update dependencies in format-check package
This commit updates the dependencies in the format-check package. The following dependencies have been added:
- @types/node @^20.8.2
- node-fetch @^3.3.2

The commit also includes updates to the package-lock.json file for these dependencies.

The pull-request-service.ts file has been modified to include an import statement for the node-fetch package. This allows the service to use the fetch function provided by node-fetch.